Исправил ошибки
